requirements.txt
fastapi
uvicorn
opencv-python
pillow
python-multipart
google-generativeai
pytesseract
jinja2